+++ src/views/controls/tab.js
@@ -1,18 +1,32 @@
 /* eslint-env browser */
 import usage from './tab.usage.js';
 
-export default function(host) {
-    host.view.define('tab', function(el, config, data, context) {
-        const {
-            content,
-            active = false,
-            disabled = false,
-            onClick,
-            value,
-            text = String(value).replace(/^./, m => m.toUpperCase())
-        } = config;
+const props = `is not array? | {
+    value,
+    text: #.props has no 'content' ? is (string or number or boolean) ?,
+    content: undefined,
+    active: undefined,
+    disabled: undefined,
+    onClick: undefined
+} | overrideProps() | {
+    ...,
+    text: text is not undefined ? text : value,
+    active is truthy,
+    disabled is truthy
+}`;
 
-        if (host.query(disabled, data, context)) {
+export default function(host) {
+    host.view.define('tab', function(el, props, data, context) {
+        const {
+            value,
+            text,
+            content,
+            active,
+            disabled,
+            onClick
+        } = props;
+
+        if (disabled) {
             el.classList.add('disabled');
         } else if (typeof onClick === 'function') {
             el.addEventListener('click', () => onClick(value));
@@ -26,7 +40,7 @@ export default function(host) {
         if (content) {
             return host.view.render(el, content, data, context);
         } else {
-            el.textContent = text;
+            el.textContent = String(text);
         }
-    }, { usage });
+    }, { usage, props });
 }