+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## next
+
+- Changed views to no longer treat some properties as a query:
+    - `checked` in `checkbox` view
+    - `expanded` in `expand` view
+    - `disabled` in `tab` view
+    - `tabs` in `tabs` view
+- Changed views to normalize props for render: `checkbox`, `expand`, `tabs` and `tab`
+
 ## 1.0.0-beta.93 (06-02-2025)
 
 - Added `navbuttons.uploadFromClipboard` button